Refactor StreamRegistry contract implementation and move instance handling to a new file.
