+++ core/contracts/river/stream_registry_contract_ex.go
@@ -1,146 +1,13 @@
 package river
 
 import (
-	"context"
-	"math/big"
-
 	"github.com/ethereum/go-ethereum/accounts/abi"
-	bind2 "github.com/ethereum/go-ethereum/accounts/abi/bind/v2"
 	"github.com/ethereum/go-ethereum/common"
 
-	"github.com/towns-protocol/towns/core/blockchain"
 	. "github.com/towns-protocol/towns/core/node/base"
 	. "github.com/towns-protocol/towns/core/node/protocol"
 	. "github.com/towns-protocol/towns/core/node/shared"
 )
-
-type StreamRegistryInstance struct {
-	*bind2.BoundContract
-	contract *StreamRegistryV1
-}
-
-func (r *StreamRegistryV1) ABI() *abi.ABI {
-	return &r.abi
-}
-
-func (r *StreamRegistryV1) NewInstance(
-	backend bind2.ContractBackend,
-	addr common.Address,
-) *StreamRegistryInstance {
-	return &StreamRegistryInstance{
-		BoundContract: bind2.NewBoundContract(addr, r.abi, backend, backend, backend),
-		contract:      r,
-	}
-}
-
-func (c *StreamRegistryInstance) GetStreamOnBlock(
-	ctx context.Context,
-	streamId StreamId,
-	blockNum blockchain.BlockNumber,
-) (*Stream, error) {
-	return blockchain.CallPtr(
-		c.BoundContract,
-		ctx,
-		"GetStream",
-		blockNum,
-		func() ([]byte, error) { return c.contract.TryPackGetStream(streamId) },
-		c.contract.UnpackGetStream,
-	)
-}
-
-func (c *StreamRegistryInstance) GetStreamOnLatestBlock(
-	ctx context.Context,
-	streamId StreamId,
-) (*Stream, error) {
-	return c.GetStreamOnBlock(ctx, streamId, 0)
-}
-
-// GetStreamWithGenesis returns stream, genesis miniblock hash, genesis miniblock, error
-func (c *StreamRegistryInstance) GetStreamWithGenesis(
-	ctx context.Context,
-	streamId StreamId,
-	blockNum blockchain.BlockNumber,
-) (*Stream, common.Hash, []byte, error) {
-	result, err := blockchain.CallPtr(
-		c.BoundContract,
-		ctx,
-		"GetStreamWithGenesis",
-		blockNum,
-		func() ([]byte, error) { return c.contract.TryPackGetStreamWithGenesis(streamId) },
-		c.contract.UnpackGetStreamWithGenesis,
-	)
-	if err != nil {
-		return nil, common.Hash{}, nil, err
-	}
-	return &result.Arg0, result.Arg1, result.Arg2, nil
-}
-
-func (c *StreamRegistryInstance) GetStreamCount(ctx context.Context, blockNum blockchain.BlockNumber) (int64, error) {
-	return blockchain.CallInt64Raw(
-		c.BoundContract,
-		ctx,
-		"GetStreamCount",
-		blockNum,
-		c.contract.PackGetStreamCount(),
-		c.contract.UnpackGetStreamCount,
-	)
-}
-
-func (c *StreamRegistryInstance) GetStreamCountOnNode(
-	ctx context.Context,
-	blockNum blockchain.BlockNumber,
-	node common.Address,
-) (int64, error) {
-	return blockchain.CallInt64(
-		c.BoundContract,
-		ctx,
-		"GetStreamCountOnNode",
-		blockNum,
-		func() ([]byte, error) { return c.contract.TryPackGetStreamCountOnNode(node) },
-		c.contract.UnpackGetStreamCountOnNode,
-	)
-}
-
-func (c *StreamRegistryInstance) GetPaginatedStreamsOnNode(
-	ctx context.Context,
-	blockNum blockchain.BlockNumber,
-	node common.Address,
-	start int64,
-	end int64,
-) ([]StreamWithId, error) {
-	return blockchain.CallValue(
-		c.BoundContract,
-		ctx,
-		"GetPaginatedStreamsOnNode",
-		blockNum,
-		func() ([]byte, error) {
-			return c.contract.TryPackGetPaginatedStreamsOnNode(node, big.NewInt(start), big.NewInt(end))
-		},
-		c.contract.UnpackGetPaginatedStreamsOnNode,
-	)
-}
-
-func (c *StreamRegistryInstance) GetPaginatedStreams(
-	ctx context.Context,
-	blockNum blockchain.BlockNumber,
-	start int64,
-	end int64,
-) ([]StreamWithId, bool, error) {
-	ret, err := blockchain.CallValue(
-		c.BoundContract,
-		ctx,
-		"GetPaginatedStreams",
-		blockNum,
-		func() ([]byte, error) {
-			return c.contract.TryPackGetPaginatedStreams(big.NewInt(start), big.NewInt(end))
-		},
-		c.contract.UnpackGetPaginatedStreams,
-	)
-	if err != nil {
-		return nil, false, err
-	}
-	return ret.Arg0, ret.Arg1, nil
-}
 
 const StreamFlagSealed uint64 = 0x1
 
@@ -213,8 +80,8 @@ func (s *StreamWithId) Nodes() []common.Address {
 type (
 	// StreamUpdated is the unified event emitted by the stream registry when a stream mutation occurs.
 	// Either when its created or modified.
-	StreamUpdated                   = StreamRegistryV1StreamUpdated
-	StreamLastMiniblockUpdateFailed = StreamRegistryV1StreamLastMiniblockUpdateFailed
+	StreamUpdated                   = StreamRegistryContractStreamUpdated
+	StreamLastMiniblockUpdateFailed = StreamRegistryContractStreamLastMiniblockUpdateFailed
 
 	// StreamUpdatedEventType defines Solidity IStreamRegistryBase.StreamEventType enum type.
 	StreamUpdatedEventType uint8
@@ -312,7 +179,7 @@ func MiniblockRefFromContractRecord(stream *Stream) *MiniblockRef {
 
 // ParseStreamUpdatedEvent parses the given stream update into the stream registry state after the update.
 // The returned slice contains one or more *StreamState or *StreamMiniblockUpdate instances.
-func ParseStreamUpdatedEvent(event *StreamRegistryV1StreamUpdated) ([]StreamUpdatedEvent, error) {
+func ParseStreamUpdatedEvent(event *StreamRegistryContractStreamUpdated) ([]StreamUpdatedEvent, error) {
 	reason := StreamUpdatedEventType(event.EventType)
 	switch reason {
 	case StreamUpdatedEventTypeAllocate:
@@ -357,7 +224,7 @@ func parseStreamIDAndStreamFromSolABIEncoded(data []byte) (*StreamWithId, error)
 	return ret, nil
 }
 
-func parseSetMiniblocksFromSolABIEncoded(event *StreamRegistryV1StreamUpdated) ([]StreamUpdatedEvent, error) {
+func parseSetMiniblocksFromSolABIEncoded(event *StreamRegistryContractStreamUpdated) ([]StreamUpdatedEvent, error) {
 	values := abi.Arguments{{Type: abi.Type{T: abi.SliceTy, Elem: &setMiniblockABIType}}}
 	unpacked, err := values.Unpack(event.Data)
 	if err != nil {
